+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+### NEXT_VERSION_TYPE=MINOR
 ### NEXT_VERSION_DESCRIPTION_BEGIN
+* Token based authentication added (https://grafana.com/docs/grafana/latest/http_api/auth/)
+* Username and password authentication deprecated
 ### NEXT_VERSION_DESCRIPTION_END
 ## [5.0.0](https://github.com/yoomoney-gradle-plugins/grafana-dashboard-plugin/pull/2) (30-03-2021)